renamed to `AutoEvent`, minimized code, added documentation
