+++ Package/Runtime/Events/AutoEvent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6b6a60c9229ba5643b589ed63d60f72a
+guid: 79c13668e4d1b0849ba03a266fd42c5a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
